requirements.txt
discord.py
motor~=2.1.0
pymongo~=3.10.1
requests~=2.23.0